allow user to ask for FQDN as public hostname (SOC-9616)
Currently CrowbarHelper.get_host_for_public_url() indiscriminately
return the IP instead of FQDN in the situation where the
node's public_name is not set and SSL is not being used.
This behavior was introduced in

https://github.com/crowbar/barclamp-glance/commit/22d3c190f1edc6f4d278982eb58076acdde16505

to address certain situations where the crowbar generated hostnames are not
resolveable externally. For SSL, we need the FQDN or hostname for
certificate validation. However, for WebSSO, certain identity providers
(i.e. Google) requires FQDN to match the authorized domain, regardless of
whether SSL is used.

In theory, we could change the behavior of get_host_for_public_url() to
always return the FQDN instead of IP as we *should be* supporting
external name resolution by now, but doing so may risk breaking
backward compatibility. Therefore, this patch introduce an option to allow
users to ask for FQDN regardless whether SSL is used without breaking
backward compatibility.