Designate: Don't add the admin node to the public network (SOC-10658)
The DNS barclamp fails to apply when designate is enabled. This is
because in that case we try and add the admin node to the public
network.
In our CI this fails as the admin node only has 1 NIC and the admin
network has a lower router preference then public, so the default route
is killed.

Due to the topology supported by designate and potential security
concerns it's better that we never connect the admin node.

There is a crowbar-openstack PR to filter the admin node out in the
designate barclamp. This is the core side.