Make sure the crowbar migrations are OK (SOC-6849)
Migrations are executed at the end of admin upgrade step.
If there was some failure, there's a big chance services step would fail
because correct chef templates might not be rendered.

So do the extra check right at the start of services step. Not during
admin upgrade step, beacuse that one is not repeatable after the reboot.