crowbar: add crowbar-pacemaker dependency (SOC-10986)
Without this dependency, barclamp application will fail with a

  uninitialized constant Chef::Recipe::CrowbarPacemakerHelper

for any node that does not have other barclamps with this
dependency declaration applied. In practice, this problem only
becomes apparent for nodes that have no roles from crowbar-openstack
applied since all Openstack barclamps declare a dependency on
crowbar-pacemaker. This would happen when removing
nova-compute-xen roles before the upgrade and not assigning
any other roles to the nodes in question. If that happens, the
upgrade will get stuck during the `services` step where role
assignments cannot be changed due to the upgrade lockout.