upgrade: Fix systemd unit listing (trivial)
If there was some file matching openstack-* pattern in current directory
the patern was expanded on execution and `systemctl list-units openstack-*`
became `systemctl list-units openstack-the-rest-of-file-name` which didn't
give the result which was expected. Quoting the pattern disabled shell
expansion and pass the argument as-is to systemctl.

(cherry picked from commit 177dca5dbab2bb443723f8ec9d54d1e973685f43)