upgrade: Make a check for SLES product version (SOC-3089)
If some packages are incorrectly upgaded during admin server upgrade,
it could blow to our faces later. Better to make some explicit check
if things went well when it's still time and check for the installed
SLES product version.

It is not necessary to check for Cloud product version because another
check for the state of crowbar migrations should cover it.

Do the extra check right at the start of services step. Not during
admin upgrade step, beacuse that one is not repeatable after the reboot.