apache2: Restart after enabling SSL flag (SOC-11029)
When SSL flag is enabled in apache's sysconfig file it is translated
into a -DSSL command argument that will be used the next time apache
is (re)started. When apache is reloaded, it is just signaled with
SIGUSR1 but not completely restarted.
Without a restart, the main process doesn't get the -DSSL option and
this causes all configs with <IfDefine SSL> to be skipped.
Additional restart in the mod_ssl recipe after enabling SSL flag is
required to handle this properly.