Add lb-mgmt-net to network.json (SOC-10904)
This change adds a network to `network.json` for configuring a
provider network for Octavia management network traffic.

Move default VLAN for fixed network to `700` so that it is the highest
VLAN id used. The Neutron barclamp calculates the ml2 vlan range as
`[vlan_fixed, vlan_fixed + num_vlans]` and we don't want to overlap
with other VLANs.

Add external router to network definition (SOC-10904)

Signed-off-by: Nicolas Bock <nicolas.bock@suse.com>