upgrade: Run DHCP evacuation (SOC-11046)
DHCP agent handles networks and ports in a way similar to L3 agent's
routers. The assignment of these to the network nodes need to be
properly orchestrated to keep things under control in upgrade scenario
where we need to move network agents out from node before upgrading
to make the process non-disruptive.