upgrade: Do not call neutron-evacuate-lbaasv2-agent with use_crm (SOC-11171)
use_crm option is there for the cases when lbaas agents are handled
by pacemaker resources. So the script puts the node into the maintenance
mode for the time it is restarting the agent.

However since SOC8 we do not use pacameker for lbaas agents, so there
is no need for maintenance mode and no need to execute the script with
use_crm option.

(cherry picked from commit 08688a2106af28db16a6b4a4a4f952fa428950fc)