Revert: Add lb-mgmt-net to network.json (SOC-10904)
The octavia network.json network is not mandatory (customers
may choose to deploy their own provider network). If we add
it to network.json, the network will always be detected and
used by Octavia.

Instead, we need to document this as an optional network
that customers may choose to add to their deployment.

This reverts commit acff1513fec04b44a4781554e9d0143074c2404d.