provisioner: check for client_user (SOC-11389)
This commit adds a check for the presence of the client_user
attribute in the Crowbar barclamp's data bag. This attribute
will only become available in chef once chef-client has run on
the Crowbar admin node or the Crowbar data bag has been
applied. Until that has happened, the code using it would
error out on all other nodes.